a1/microservices/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
jsonpointer==2.1
requests-oauthlib==1.4.0
mysql-connector-python==8.3.0
requests==2.31.0
bcrypt==4.1.2
python-dotenv==0.19.1
pika==1.3.2
pytz